feat: add base request handler
body: this code will still be non-functional and will need a
specific request handler.  This handler receives either a dict
or yaml file which describes the type of request (GET, PUT), the
tables to insert or receive data from, and the specific filters.
This will act much like a REST api, but will not require an http
request.  These request handlers could be modified fairly easily
to transform them into actual REST apis.